session/requirements/requirements.txt
Django==5.2.2
django-ipware==7.0.1
djangorestframework==3.16.0
geoip2==5.1.0
python-decouple==3.8
django-recaptcha==4.1.0
django-jazzmin==3.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-json-logger==3.3.0
django-redis==6.0.0
redis==5.2.0
celery==5.5.3
flower==2.0.1
watchfiles==1.1.0
django-celery-beat==2.8.1
django-celery-email==3.0.0
user-agents==2.2.0
qrcode==8.2
pyotp==2.9.0
